refactor: ktlint 포맷팅
