+++ app/src/main/java/woowacourse/shopping/data/CartRepositoryImpl.kt
@@ -1,18 +1,17 @@
 package woowacourse.shopping.data
 
+import woowacourse.shopping.domain.CartRepository
 import woowacourse.shopping.model.Product
 
 // TODO: Step2 - CartProductDao를 참조하도록 변경
-class CartRepositoryImpl: CartRepository {
-
+class CartRepositoryImpl : CartRepository {
     private val cartProducts: MutableList<Product> = mutableListOf()
+
     override fun addCartProduct(product: Product) {
         cartProducts.add(product)
     }
 
-    override fun getAllCartProducts(): List<Product> {
-        return cartProducts.toList()
-    }
+    override fun getAllCartProducts(): List<Product> = cartProducts.toList()
 
     override fun deleteCartProduct(id: Int) {
         cartProducts.removeAt(id)
