+++ app/src/main/java/woowacourse/shopping/ui/cart/CartProductAdapter.kt
@@ -9,7 +9,6 @@ class CartProductAdapter(
     onClickDelete: (position: Int) -> Unit,
     private val dateFormatter: DateFormatter,
 ) : RecyclerView.Adapter<CartProductViewHolder>() {
-
     private val items: MutableList<Product> = items.toMutableList()
 
     private val onClickDelete = { position: Int ->
@@ -17,11 +16,15 @@ class CartProductAdapter(
         removeItem(position)
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CartProductViewHolder {
-        return CartProductViewHolder.from(parent, dateFormatter, onClickDelete)
-    }
+    override fun onCreateViewHolder(
+        parent: ViewGroup,
+        viewType: Int,
+    ): CartProductViewHolder = CartProductViewHolder.from(parent, dateFormatter, onClickDelete)
 
-    override fun onBindViewHolder(holder: CartProductViewHolder, position: Int) {
+    override fun onBindViewHolder(
+        holder: CartProductViewHolder,
+        position: Int,
+    ) {
         holder.bind(items[position])
     }
